Conflicts:
	.gitignore
	application/bundles.php
	application/config/application.php
	application/config/session.php
	application/config/strings.php
	application/routes.php
	application/views/home/index.php
	readme.md

+++ laravel/file.php
@@ -1,4 +1,4 @@
-<?php namespace Laravel; use Closure, FilesystemIterator as fIterator;
+<?php namespace Laravel; use FilesystemIterator as fIterator;
 
 class File {
 
@@ -61,16 +61,40 @@ class File {
 	 * Delete a file.
 	 *
 	 * @param  string  $path
-	 * @return void
+	 * @return bool
 	 */
 	public static function delete($path)
 	{
-		if (static::exists($path)) @unlink($path);
+		if (static::exists($path)) return @unlink($path);
+	}
+
+	/**
+	 * Move a file to a new location.
+	 *
+	 * @param  string  $path
+	 * @param  string  $target
+	 * @return void
+	 */
+	public static function move($path, $target)
+	{
+		return rename($path, $target);
+	}
+
+	/**
+	 * Copy a file to a new location.
+	 *
+	 * @param  string  $path
+	 * @param  string  $target
+	 * @return void
+	 */
+	public static function copy($path, $target)
+	{
+		return copy($path, $target);
 	}
 
 	/**
 	 * Extract the file extension from a file path.
-	 * 
+	 *
 	 * @param  string  $path
 	 * @return string
 	 */
@@ -161,8 +185,7 @@ class File {
 
 		// The MIME configuration file contains an array of file extensions and
 		// their associated MIME types. We will spin through each extension the
-		// developer wants to check to determine if the file's MIME type is in
-		// the list of MIMEs we have for that extension.
+		// developer wants to check and look for the MIME type.
 		foreach ((array) $extensions as $extension)
 		{
 			if (isset($mimes[$extension]) and in_array($mime, (array) $mimes[$extension]))
@@ -175,6 +198,18 @@ class File {
 	}
 
 	/**
+	 * Create a new directory.
+	 *
+	 * @param  string  $path
+	 * @param  int     $chmod
+	 * @return void
+	 */
+	public static function mkdir($path, $chmod = 0777)
+	{
+		return ( ! is_dir($path)) ? mkdir($path, $chmod, true) : true;
+	}
+
+	/**
 	 * Move a directory from one location to another.
 	 *
 	 * @param  string  $source
@@ -184,7 +219,7 @@ class File {
 	 */
 	public static function mvdir($source, $destination, $options = fIterator::SKIP_DOTS)
 	{
-		static::cpdir($source, $destination, true, $options);
+		return static::cpdir($source, $destination, true, $options);
 	}
 
 	/**
@@ -198,7 +233,7 @@ class File {
 	 */
 	public static function cpdir($source, $destination, $delete = false, $options = fIterator::SKIP_DOTS)
 	{
-		if ( ! is_dir($source)) return;
+		if ( ! is_dir($source)) return false;
 
 		// First we need to create the destination directory if it doesn't
 		// already exists. This directory hosts all of the assets we copy
@@ -222,7 +257,7 @@ class File {
 			{
 				$path = $item->getRealPath();
 
-				static::cpdir($path, $location, $delete, $options);
+				if (! static::cpdir($path, $location, $delete, $options)) return false;
 
 				if ($delete) @rmdir($item->getRealPath());
 			}
@@ -232,22 +267,26 @@ class File {
 			// files with the same name.
 			else
 			{
-				copy($item->getRealPath(), $location);
+				if(! copy($item->getRealPath(), $location)) return false;
 
 				if ($delete) @unlink($item->getRealPath());
 			}
 		}
 
-		if ($delete) rmdir($source);
+		unset($items);
+		if ($delete) @rmdir($source);
+
+		return true;
 	}
 
 	/**
 	 * Recursively delete a directory.
 	 *
 	 * @param  string  $directory
+	 * @param  bool    $preserve
 	 * @return void
 	 */
-	public static function rmdir($directory)
+	public static function rmdir($directory, $preserve = false)
 	{
 		if ( ! is_dir($directory)) return;
 
@@ -268,7 +307,19 @@ class File {
 			}
 		}
 
-		@rmdir($directory);
+		unset($items);
+		if ( ! $preserve) @rmdir($directory);
+	}
+
+	/**
+	 * Empty the specified directory of all files and folders.
+	 *
+	 * @param  string  $directory
+	 * @return void
+	 */
+	public static function cleandir($directory)
+	{
+		return static::rmdir($directory, true);
 	}
 
 	/**
@@ -286,8 +337,7 @@ class File {
 
 		// To get the latest created file, we'll simply spin through the
 		// directory, setting the latest file if we encounter a file
-		// with a UNIX timestamp greater than the latest one we
-		// have encountered thus far in the loop.
+		// with a UNIX timestamp greater than the latest one.
 		foreach ($items as $item)
 		{
 			if ($item->getMTime() > $time) $latest = $item;
